103-fibonacci C program

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -5,10 +5,10 @@
  */
 int main(void)
 {
-	long i = 0;
-	long j = 1;
-	long sum1;
-	long sum2;
+	int i = 0;
+	int j = 1;
+	int sum1;
+	int sum2;
 
 	while (sum1 < 4000000)
 	{
@@ -18,6 +18,6 @@ int main(void)
 		if (sum1 % 2 == 0)
 			sum2 = sum2 + sum1;
 	}
-	printf("%lu", sum2);
+	printf("%d", sum2);
 return (0);
 }
